Rewire component library creation into chat action commands
